Optimisation! xsge
This is a big one, as Alex sent me clm.tex (not tracked at
this point), which has over 9000 lines, and around 12000 objects.

BEFORE
real   5m45.170s
user    5m42.289s
sys 0m1.144s

AFTER
real    0m50.810s
user    0m48.147s
sys 0m1.324s

The crux of the changes in this commit are to change the substitutions
to be *global* using xseg.

Unfortunately, this has negatively affected one of the easier bench mark
test cases in test-cases/benchmarks/sampleBEFORE.tex, but compared
to the results above, it's not worth worrying about.