pre-commit: Add `--overwriteIfDifferent --silent --local` to default args
Without `--overwriteIfDifferent`, the pre-commit hook does not work at
all. Output is printed to stdout but never seen by the user. Note that
this still spams useless backup files in the git repository (the git
staging area already acts as backup).

The `--silent` flag avoid spamming the terminal in case the hook
modifies any files, since the output is not that helpful anyways.

The `--local` flag is useful because you usually want every contributor
to use the same settings, so the local flags allows putting a
`.latexindent.yaml` in the root of the git repository.